// frontend/packages/zobi-ui-chart-controls/src/types.ts
export interface StandardizedControls {
  metrics: QueryFormMetric[];
  columns: QueryFormColumn[];
}